requirements.txt
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bitsandbytes==0.49.1
blinker==1.9.0
blis==1.3.3
cachetools==6.2.6
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
datasets==4.5.0
dill==0.4.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
faiss-cpu==1.13.2
filelock==3.20.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
narwhals==2.16.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.5
packaging==26.0
pandas==2.3.3
peft==0.18.1
pillow==12.0.0
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
torch==2.10.0+cu130
torchaudio==2.10.0+cu130
torchvision==0.25.0+cu130
tornado==6.5.4
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.3
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
